pyproject.toml
[project]
name = "nohayonline_scraper"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = ["scrapy>=2.11", "psycopg[binary]>=3.2", "python-dotenv>=1.0"]
